requirements.txt
asgiref==3.7.2
Django==4.2.7
django-admin-interface==0.26.1
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-cors-headers==4.3.1
django-js-asset==2.1.0
django-session-timeout==0.1.0
django-tinymce==3.6.1
Pillow==10.1.0
python-slugify==8.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
